requirements.txt
google==3.0.0
python-dotenv==1.1.0
google-generativeai==0.8.5

